#110 [General Master Data][Permission]

+++ app/Http/Controllers/Master/General/BroadbandPacketController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\BroadbandPacketRequest;
 use App\Models\BroadbandPacket;
 use App\Service\Master\BroadbandPacketService;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\View\View;
@@ -22,13 +23,21 @@ class BroadbandPacketController extends Controller
         $this->broadbandPacketService = new BroadbandPacketService();
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function index(): View
     {
+        $this->authorize('view', BroadbandPacket::class);
         return view('pages.general-master-data.broadband-packet.index');
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function search(Request $request): JsonResponse
     {
+        $this->authorize('view', BroadbandPacket::class);
         $search = $request->input('search');
         $query = $this->broadbandPacket->data($request);
 
@@ -42,13 +51,21 @@ class BroadbandPacketController extends Controller
         return response()->json($data);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function data(Request $request): JsonResponse
     {
+        $this->authorize('view', BroadbandPacket::class);
         return response()->json($this->broadbandPacketService->data($request));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function store(BroadbandPacketRequest $request): JsonResponse
     {
+        $this->authorize('create', BroadbandPacket::class);
         $data = $request->validated();
         $data['branch_id'] = $request->user()->branch_id;
         BroadbandPacket::create($data);
@@ -58,14 +75,22 @@ class BroadbandPacketController extends Controller
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function edit(BroadbandPacket $broadbandPacket): JsonResponse
     {
+        $this->authorize('update', $broadbandPacket);
         return response()->json($broadbandPacket);
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function update(BroadbandPacketRequest $request, BroadbandPacket $broadbandPacket): JsonResponse
     {
+        $this->authorize('update', $broadbandPacket);
         $data = $request->validated();
         $data['branch_id'] = $request->user()->branch_id;
         $broadbandPacket->update($data);
@@ -75,8 +100,12 @@ class BroadbandPacketController extends Controller
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function destroy(Request $request, BroadbandPacket $broadbandPacket): JsonResponse
     {
+        $this->authorize('delete', $broadbandPacket);
         $implodeID = implode(',', $request->get('id'));
         $explodeID = explode(',', $implodeID);
         $broadbandPacket->whereIn('id', $explodeID)->delete();
